Update reservation and entity classes, add company and guest creation logic
- Added `Reservation` entity construction from `ReservationDto` with support for guest or company-based reservations.
- Implemented logic to create or find existing guests and companies in `ReservationService`.
- Changed ID types from `Long` to `Integer` across DTOs and entities to maintain consistency.
- Introduced `created` field in `Company`, `Guest`, and `ReservationActivity` entities to track creation timestamps.
- Modified return types and parameters in `ReservationController` to enhance API response clarity.
- Adjusted repository interfaces for `Company` and `Guest` to support finding by `companyName` and `email`, respectively.
- Developed service classes for `ActivityReservation`, `CompanyService`, and `GuestService` with methods to handle entity creation and updates.
- Updated `updateReservation` method in `ReservationService` to support both guest and company associations.

+++ src/main/java/dat3/adventureXP/dto/EquipmentDto.java
@@ -7,11 +7,11 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class EquipmentDto {
-    private Long id;
+    private Integer id;
     private String name;
-    private Long activityId;  // Use activity ID instead of the Activity object
+    private Integer activityId;  // Use activity ID instead of the Activity object
 
-    public EquipmentDto(Long id, String name, Long activityId) {
+    public EquipmentDto(Integer id, String name, Integer activityId) {
         this.id = id;
         this.name = name;
         this.activityId = activityId;
